fix: move environment checks to a dedicated monitor
- Move environment checks to a dedicated monitor
- Remove hugepage condition after disabling v2 data engine

Longhorn 9983
